Split tests per field
