+++ test/Form_test.js
@@ -1,0 +1,356 @@
+
+/*eslint no-unused-vars: [2, { "varsIgnorePattern": "^d$" }]*/
+
+import { expect } from "chai";
+import sinon from "sinon";
+import React from "react";
+import { Simulate, renderIntoDocument } from "react-addons-test-utils";
+import { findDOMNode } from "react-dom";
+
+import Form from "../src";
+import { createComponent } from "./test_utils";
+
+describe("Form", () => {
+  var sandbox;
+
+  beforeEach(() => {
+    sandbox = sinon.sandbox.create();
+  });
+
+  afterEach(() => {
+    sandbox.restore();
+  });
+
+  describe("Empty schema", () => {
+    it("should render a form tag", () => {
+      const {node} = createComponent({schema: {}});
+
+      expect(node.tagName).eql("FORM");
+    });
+
+    it("should render a submit button", () => {
+      const {node} = createComponent({schema: {}});
+
+      expect(node.querySelectorAll("button[type=submit]"))
+        .to.have.length.of(1);
+    });
+
+    it("should render children buttons", () => {
+      const props = {schema: {}};
+      const comp = renderIntoDocument(
+        <Form {...props}>
+          <button type="submit">Submit</button>
+          <button type="submit">Another submit</button>
+        </Form>
+      );
+      const node = findDOMNode(comp);
+      expect(node.querySelectorAll("button[type=submit]"))
+        .to.have.length.of(2);
+    });
+  });
+
+  describe("Custom submit buttons", () => {
+    it("should submit the form when clicked", () => {
+      const onSubmit = sandbox.spy();
+      const comp = renderIntoDocument(
+        <Form onSubmit={ onSubmit } schema={ {} }>
+          <button type="submit">Submit</button>
+          <button type="submit">Another submit</button>
+        </Form>
+      );
+      const node = findDOMNode(comp);
+      const buttons = node.querySelectorAll("button[type=submit]");
+      buttons[0].click();
+      buttons[1].click();
+      sinon.assert.calledTwice(onSubmit);
+    });
+  });
+
+  describe("Defaults array items default propagation", () => {
+    const schema = {
+      type: "object",
+      title: "lvl 1 obj",
+      properties: {
+        object: {
+          type: "object",
+          title: "lvl 2 obj",
+          properties: {
+            array: {
+              type: "array",
+              items: {
+                type: "object",
+                title: "lvl 3 obj",
+                properties: {
+                  bool: {
+                    type: "boolean",
+                    default: true
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+    };
+
+    it("should propagate deeply nested defaults to form state", () => {
+      const {comp, node} = createComponent({schema});
+
+      Simulate.click(node.querySelector(".array-item-add button"));
+      Simulate.submit(node);
+
+      expect(comp.state.formData).eql({
+        object: {
+          array: [
+            {
+              bool: true
+            }
+          ]
+        }
+      });
+    });
+  });
+
+  describe("Validation", () => {
+    describe("Required fields", () => {
+      const schema = {
+        type: "object",
+        required: ["foo"],
+        properties: {
+          foo: {type: "string"},
+          bar: {type: "string"},
+        }
+      };
+
+      var comp, node, onError;
+
+      beforeEach(() => {
+        onError = sandbox.spy();
+        const compInfo = createComponent({schema, formData: {
+          foo: undefined
+        }, onError});
+        comp = compInfo.comp;
+        node = compInfo.node;
+
+        Simulate.submit(node);
+      });
+
+      it("should validate a required field", () => {
+        expect(comp.state.errors)
+          .to.have.length.of(1);
+        expect(comp.state.errors[0].message)
+          .eql(`requires property "foo"`);
+      });
+
+      it("should render errors", () => {
+        expect(node.querySelectorAll(".errors li"))
+          .to.have.length.of(1);
+        expect(node.querySelector(".errors li").textContent)
+          .eql(`instance requires property "foo"`);
+      });
+
+      it("should trigger the onError handler", () => {
+        sinon.assert.calledWith(onError, sinon.match(errors => {
+          return errors[0].message === `requires property "foo"`;
+        }));
+      });
+    });
+
+    describe("Min length", () => {
+      const schema = {
+        type: "object",
+        required: ["foo"],
+        properties: {
+          foo: {
+            type: "string",
+            minLength: 10,
+          },
+        }
+      };
+
+      var comp, node, onError;
+
+      beforeEach(() => {
+        onError = sandbox.spy();
+        const compInfo = createComponent({schema, formData: {
+          foo: "123456789"
+        }, onError});
+        comp = compInfo.comp;
+        node = compInfo.node;
+
+        Simulate.submit(node);
+      });
+
+      it("should validate a minLength field", () => {
+        expect(comp.state.errors)
+          .to.have.length.of(1);
+        expect(comp.state.errors[0].message)
+          .eql(`does not meet minimum length of 10`);
+      });
+
+      it("should render errors", () => {
+        expect(node.querySelectorAll(".errors li"))
+          .to.have.length.of(1);
+        expect(node.querySelector(".errors li").textContent)
+          .eql("instance.foo does not meet minimum length of 10");
+      });
+
+      it("should trigger the onError handler", () => {
+        sinon.assert.calledWith(onError, sinon.match(errors => {
+          return errors[0].message ===
+            "does not meet minimum length of 10";
+        }));
+      });
+    });
+  });
+
+  describe("Submit handler", () => {
+    it("should call provided submit handler with form state", () => {
+      const schema = {
+        type: "object",
+        properties: {
+          foo: {type: "string"},
+        }
+      };
+      const formData = {
+        foo: "bar"
+      };
+      const onSubmit = sandbox.spy();
+      const {comp, node} = createComponent({schema, formData, onSubmit});
+
+      Simulate.submit(node);
+
+      sinon.assert.calledWithExactly(onSubmit, comp.state);
+    });
+
+    it("should not call provided submit handler on validation errors", () => {
+      const schema = {
+        type: "object",
+        properties: {
+          foo: {
+            type: "string",
+            minLength: 1,
+          },
+        }
+      };
+      const formData = {
+        foo: ""
+      };
+      const onSubmit = sandbox.spy();
+      const onError = sandbox.spy();
+      const {node} = createComponent({schema, formData, onSubmit, onError});
+
+      Simulate.submit(node);
+
+      sinon.assert.notCalled(onSubmit);
+    });
+  });
+
+  describe("Change handler", () => {
+    it("should call provided change handler on form state change", () => {
+      const schema = {
+        type: "object",
+        properties: {
+          foo: {
+            type: "string",
+          },
+        }
+      };
+      const formData = {
+        foo: ""
+      };
+      const onChange = sandbox.spy();
+      const {node} = createComponent({schema, formData, onChange});
+
+      Simulate.change(node.querySelector("[type=text]"), {
+        target: {value: "new"}
+      });
+
+      sinon.assert.calledWithMatch(onChange, {
+        formData: {
+          foo: "new"
+        }
+      });
+    });
+  });
+
+  describe("Error handler", () => {
+    it("should call provided error handler on validation errors", () => {
+      const schema = {
+        type: "object",
+        properties: {
+          foo: {
+            type: "string",
+            minLength: 1,
+          },
+        }
+      };
+      const formData = {
+        foo: ""
+      };
+      const onError = sandbox.spy();
+      const {node} = createComponent({schema, formData, onError});
+
+      Simulate.submit(node);
+
+      sinon.assert.calledOnce(onError);
+    });
+  });
+
+  describe("External formData updates", () => {
+    describe("root level", () => {
+      it("should update form state from new formData prop value", () => {
+        const schema = {type: "string"};
+        const {comp} = createComponent({schema});
+
+        comp.componentWillReceiveProps({formData: "yo"});
+
+        expect(comp.state.formData).eql("yo");
+      });
+
+      it("should validate formData when the schema is updated", () => {
+        const {comp} = createComponent({schema: {type: "string"}});
+
+        comp.componentWillReceiveProps({formData: "yo", schema: {type: "number"}});
+
+        expect(comp.state.errors).to.have.length.of(1);
+        expect(comp.state.errors[0].stack)
+          .to.eql("instance is not of a type(s) number");
+      });
+    });
+
+    describe("object level", () => {
+      it("should update form state from new formData prop value", () => {
+        const schema = {
+          type: "object",
+          properties: {
+            foo: {
+              type: "string"
+            }
+          }
+        };
+        const {comp} = createComponent({schema});
+
+        comp.componentWillReceiveProps({formData: {foo: "yo"}});
+
+        expect(comp.state.formData).eql({foo: "yo"});
+      });
+    });
+
+    describe("array level", () => {
+      it("should update form state from new formData prop value", () => {
+        const schema = {
+          type: "array",
+          items: {
+            type: "string"
+          }
+        };
+        const {comp} = createComponent({schema});
+
+        comp.componentWillReceiveProps({formData: ["yo"]});
+
+        expect(comp.state.formData).eql(["yo"]);
+      });
+    });
+  });
+});
