+++ test/SchemaField_test.js
@@ -1,0 +1,21 @@
+import React from "react";
+import { expect } from "chai";
+import SchemaField from "../src/components/fields/SchemaField";
+
+import { createComponent } from "./test_utils";
+
+describe("Custom SchemaField", () => {
+  const CustomSchemaField = function(props) {
+    return (<div id="custom"><SchemaField {...props} /></div>);
+  };
+
+  it("should use the specified custom SchemaType property", () => {
+    const {node} = createComponent({
+      schema: {type: "string"},
+      SchemaField: CustomSchemaField
+    });
+
+    expect(node.querySelectorAll("#custom > .field input[type=text]"))
+      .to.have.length.of(1);
+  });
+});
